Update README.md. Add TV support to log view.
